feat: make jump_to serve Learning MFE urls, if it's active
The /jump_to/ LMS endpoint is used in a number of places
to direct users to courseware. It currently only redirects to
Legacy courseware URLs, which then conditionally may
redirect to the Learning MFE.

Two issues with this:
1. Performance: in most cases, going to Legacy first
   is just an extra redirect.
2. Educator view: educators often land on Legacy,
   even when their students end up on the Learning MFE.

This commit makes it so that /jump_to/ directs
users to whatever the default courseware experience is
for them.

TNL-7796